+++ graph/ResultCachedGraph.h
@@ -10,15 +10,15 @@ using namespace std;
 class ResultCachedGraph : public CouplingGraph {
 private:
     shared_ptr<CouplingGraph> wrapped;
-    unordered_map<string, float> supportCache;
-    unordered_map<pair<string, string>, float, pair_hash> couplingCache;
+    unordered_map<string, double> supportCache;
+    unordered_map<pair<string, string>, double, pair_hash> couplingCache;
 
 public:
     explicit ResultCachedGraph(const shared_ptr<CouplingGraph>& wrapped);
 
     const vector<string> & getNodeSet() override;
-    float getNormalizedSupport(const string &node) override;
-    float getNormalizedCoupling(const string &a, const string &b) override;
+    double getNormalizedSupport(const string &node) override;
+    double getNormalizedCoupling(const string &a, const string &b) override;
     void printStatistics() override;
 };
